backend/requirements.txt
# Direct dependencies for Tailor Project Backend
fastapi>=0.133.0
uvicorn[standard]>=0.41.0
langgraph>=1.0.0
pydantic[email]>=2.12.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
asyncpg>=0.30.0
python-dotenv>=1.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Testing
pytest>=9.0.0
httpx>=0.28.0
pytest-asyncio>=1.3.0
aiosqlite>=0.20.0
